more paging and work

+++ Proiect_DAW/Controllers/SubjectsController.cs
@@ -4,12 +4,15 @@ using System.Linq;
 using System.Web;
 using System.Web.Mvc;
 using Proiect_DAW.Models;
+using Microsoft.AspNet.Identity;
+using PagedList;
+using PagedList.Mvc;
 
 namespace Proiect_DAW.Controllers
 {
     public class SubjectsController : Controller
     {
-        private SubjectDBContext db = new SubjectDBContext();
+        private ApplicationDbContext db = new ApplicationDbContext();
 
         public ActionResult Index()
         {
@@ -26,10 +29,16 @@ namespace Proiect_DAW.Controllers
             return View();
         }
 
-        public ActionResult Show(int id)
+        public ActionResult Show(int id, int? page)
         {
             Subject subject = db.Subjects.Find(id);
             ViewBag.Subject = subject;
+
+            int pageSize = 5;
+            int pageIndex = page.HasValue ? Convert.ToInt32(page) : 1;
+            PagedList<Reply> replies = new PagedList<Reply>(subject.Replies, pageIndex, pageSize);
+            ViewBag.Replies = replies;
+
             return View();
         }
 
@@ -45,6 +54,8 @@ namespace Proiect_DAW.Controllers
             try
             {
                 subject.Data = System.DateTime.Now.ToString();
+                subject.UserId = User.Identity.GetUserId();
+                subject.NumberOfViews = 0;
                 db.Subjects.Add(subject);
                 db.SaveChanges();
                 return RedirectToAction("Index");
@@ -59,7 +70,16 @@ namespace Proiect_DAW.Controllers
         {
             Subject subject = db.Subjects.Find(id);
             ViewBag.Subject = subject;
-            return View();
+
+            if (subject.UserId == User.Identity.GetUserId() || User.IsInRole("Editor") || User.IsInRole("Administrator"))
+            {
+                return View();
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui articol care nu va apartine!";
+                return RedirectToAction("Index"); //TODO
+            }
         }
 
         [HttpPut]
@@ -68,14 +88,22 @@ namespace Proiect_DAW.Controllers
             try
             {
                 Subject subject = db.Subjects.Find(id);
-                if (TryUpdateModel(subject))
+                if (subject.UserId == User.Identity.GetUserId() || User.IsInRole("Editor") || User.IsInRole("Administrator"))
                 {
-                    subject.Title = requestSubject.Title;
-                    subject.Content = requestSubject.Content;
-                    subject.Data = System.DateTime.Now.ToString();
-                    db.SaveChanges();
+                    if (TryUpdateModel(subject))
+                    {
+                        subject.Title = requestSubject.Title;
+                        subject.Content = requestSubject.Content;
+                        subject.Data = System.DateTime.Now.ToString();
+                        db.SaveChanges();
+                    }
+                    return RedirectToAction("Index");
                 }
-                return RedirectToAction("Index");
+                else
+                {
+                    TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui articol care nu va apartine!";
+                    return RedirectToAction("Index");
+                }
             }
             catch (Exception e)
             {
@@ -87,9 +115,17 @@ namespace Proiect_DAW.Controllers
         public ActionResult Delete(int id)
         {
             Subject subject = db.Subjects.Find(id);
-            db.Subjects.Remove(subject);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (subject.UserId == User.Identity.GetUserId() || User.IsInRole("Editor") || User.IsInRole("Administrator"))
+            {
+                db.Subjects.Remove(subject);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui articol care nu va apartine!";
+                return RedirectToAction("Index");
+            }
         }
     }
 }
